Add CollectionNameResolver tests and improve documentation
The exception thrown when a type is not mapped to a collection has been renamed from `MissingDbCollectionConfiguration` to `MissingDbCollectionConfigurationException` for clearer communication of its purpose. The terms used in `ICollectionNameResolver` interface methods have also been clarified.

Furthermore, logic has been added in `CollectionNameResolver.cs` to throw an `ArgumentException` when attempting to assign an empty string as a collection name to a type.